# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  user: "user",
  admin: "admin",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  createdAt: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface ParkingLocation {
  id: number;
  name: string;
  address: string;
  lat: number;
  lng: number;
  totalSlots: number;
  availableSlots: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type ParkingSlotType =
  (typeof ParkingSlotType)[keyof typeof ParkingSlotType];

export const ParkingSlotType = {
  standard: "standard",
  handicap: "handicap",
  electric: "electric",
  compact: "compact",
} as const;

export type ParkingSlotStatus =
  (typeof ParkingSlotStatus)[keyof typeof ParkingSlotStatus];

export const ParkingSlotStatus = {
  available: "available",
  booked: "booked",
  maintenance: "maintenance",
} as const;

export interface ParkingSlot {
  id: number;
  locationId: number;
  slotNumber: string;
  floor: string;
  type: ParkingSlotType;
  status: ParkingSlotStatus;
  pricePerHour: number;
  /** @nullable */
  bookedBy?: string | null;
  /** @nullable */
  bookedUntil?: string | null;
  /** @nullable */
  locationName?: string | null;
}

export interface CreateBookingBody {
  slotId: number;
  hours: number;
  vehiclePlate: string;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  active: "active",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface Booking {
  id: number;
  userId: number;
  slotId: number;
  vehiclePlate: string;
  startTime: string;
  endTime: string;
  hours: number;
  totalCost: number;
  status: BookingStatus;
  createdAt: string;
  /** @nullable */
  slotNumber?: string | null;
  /** @nullable */
  floor?: string | null;
  /** @nullable */
  locationName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
}

export interface DashboardSummary {
  totalSlots: number;
  availableSlots: number;
  bookedSlots: number;
  maintenanceSlots: number;
  totalLocations: number;
  activeBookings: number;
  todayRevenue: number;
}

export type AdminCreateSlotBodyType =
  (typeof AdminCreateSlotBodyType)[keyof typeof AdminCreateSlotBodyType];

export const AdminCreateSlotBodyType = {
  standard: "standard",
  handicap: "handicap",
  electric: "electric",
  compact: "compact",
} as const;

export interface AdminCreateSlotBody {
  locationId: number;
  slotNumber: string;
  floor: string;
  type: AdminCreateSlotBodyType;
  pricePerHour: number;
}

export type AdminUpdateSlotBodyType =
  (typeof AdminUpdateSlotBodyType)[keyof typeof AdminUpdateSlotBodyType];

export const AdminUpdateSlotBodyType = {
  standard: "standard",
  handicap: "handicap",
  electric: "electric",
  compact: "compact",
} as const;

export type AdminUpdateSlotBodyStatus =
  (typeof AdminUpdateSlotBodyStatus)[keyof typeof AdminUpdateSlotBodyStatus];

export const AdminUpdateSlotBodyStatus = {
  available: "available",
  booked: "booked",
  maintenance: "maintenance",
} as const;

export interface AdminUpdateSlotBody {
  slotNumber?: string;
  floor?: string;
  type?: AdminUpdateSlotBodyType;
  status?: AdminUpdateSlotBodyStatus;
  pricePerHour?: number;
}

export type GetSlotsParams = {
  /**
   * Filter by location ID
   * @nullable
   */
  locationId?: number | null;
};