# lib/api-zod/src/generated/types/adminCreateSlotBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Parking Finder API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminCreateSlotBodyType } from "./adminCreateSlotBodyType";

export interface AdminCreateSlotBody {
  locationId: number;
  slotNumber: string;
  floor: string;
  type: AdminCreateSlotBodyType;
  pricePerHour: number;
}